+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Fixed a crash while finishing export when using back- or foreground layers.
 * Fixed export when using a custom procedure containing a Color argument.
+* Fixed export of layers with locks (position, alpha channel, etc.). All locks are now ignored during export.
 * Removed the .exe installer for Windows.